requirements/dev.txt
-r base.txt
django-debug-toolbar==4.4.6
omni-authify # added by Mukhsin 30.03.2025
django-filter==25.1 # added from Oybek 29.04.2025
